Moved test class to its own file.
